Ajuste BD + Ajuste API

+++ ConnectFit-BackEnd/database/migrations/2023_10_08_161317_create_area_medida_corporal_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('area_medida_corporais', function (Blueprint $table) {
-            $table->unsignedTinyInteger('idArea');
+            $table->unsignedInteger('idArea');
             $table->unsignedInteger('idMedida');
             $table->timestamps();
